Implementable rollable formulas for powers
Both the 'attribute' & 'target' fields of a power have the new 'Roll Value' and 'Roll Formula' fields. Filling anything in the 'Roll Formula' field adds a 'Roll' button to the character sheet. This sheet will roll all dice formulas on a given ability at once, and update the 'roll value' field accordingly. This field can be referenced in the 'value' field with '@roll'.